+++ src/app/services/sdk/messages/chat.phone_pb.d.ts
@@ -989,6 +989,26 @@ export namespace PhoneActionPicked {
   }
 }
 
+export class PhoneActionRestarted extends jspb.Message {
+  getSender(): boolean;
+  setSender(value: boolean): void;
+
+  serializeBinary(): Uint8Array;
+  toObject(includeInstance?: boolean): PhoneActionRestarted.AsObject;
+  static toObject(includeInstance: boolean, msg: PhoneActionRestarted): PhoneActionRestarted.AsObject;
+  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
+  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
+  static serializeBinaryToWriter(message: PhoneActionRestarted, writer: jspb.BinaryWriter): void;
+  static deserializeBinary(bytes: Uint8Array): PhoneActionRestarted;
+  static deserializeBinaryFromReader(message: PhoneActionRestarted, reader: jspb.BinaryReader): PhoneActionRestarted;
+}
+
+export namespace PhoneActionRestarted {
+  export type AsObject = {
+    sender: boolean,
+  }
+}
+
 export class PhoneActionMediaSettingsUpdated extends jspb.Message {
   getVideo(): boolean;
   setVideo(value: boolean): void;
@@ -1153,7 +1173,7 @@ export enum PhoneCallAction {
   PHONECALLADMINUPDATED = 14,
   PHONECALLSCREENSHARE = 15,
   PHONECALLPICKED = 16,
-  PHONECALLRESERVED13 = 17,
+  PHONECALLRESTARTED = 17,
   PHONECALLRESERVED14 = 18,
   PHONECALLRESERVED15 = 19,
   PHONECALLRESERVED16 = 20,