+++ src/index.tsx
@@ -14,9 +14,9 @@ import * as serviceWorkerRegistration from "./serviceWorkerRegistration";
 
 import './index.css';
 
-export const C_VERSION = '1.1.18';
+export const C_VERSION = '1.1.16';
 export const C_ELECTRON_VERSIONS = ['12.0.2'];
-export const C_APP_VERSION = '0.33.0';
+export const C_APP_VERSION = '0.34.0';
 export const isProd = (!process || !process.env || process.env.NODE_ENV !== 'development');
 
 const App = React.lazy(() => import('./App'));